@@ -3,18 +3,17 @@ package googlesheets
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
 	"os"
 
 	"github.com/mitchellh/go-homedir"
-	"github.com/turbot/steampipe-plugin-sdk/v3/grpc/proto"
-	"github.com/turbot/steampipe-plugin-sdk/v3/plugin"
+	"github.com/turbot/steampipe-plugin-sdk/v5/grpc/proto"
+	"github.com/turbot/steampipe-plugin-sdk/v5/plugin"
 )
 
 // Returns the ID of the current working spreadsheet
-func getSpreadsheetID(_ context.Context, p *plugin.Plugin) string {
+func getSpreadsheetID(ctx context.Context, p *plugin.TableMapData) string {
 	spreadsheetCacheKey := "googlesheets.spreadsheetID"
-	if cachedData, ok := p.ConnectionManager.Cache.Get(spreadsheetCacheKey); ok {
+	if cachedData, ok := p.ConnectionCache.Get(ctx, spreadsheetCacheKey); ok {
 		return cachedData.(string)
 	}
 
@@ -25,7 +24,36 @@ func getSpreadsheetID(_ context.Context, p *plugin.Plugin) string {
 		spreadsheetID = *googleSheetsConfig.SpreadsheetId
 	}
 
-	p.ConnectionManager.Cache.Set(spreadsheetCacheKey, spreadsheetID)
+	err := p.ConnectionCache.Set(ctx, spreadsheetCacheKey, spreadsheetID)
+
+	if err != nil {
+		plugin.Logger(ctx).Error("getSpreadsheetID", "connection set error", err)
+		return err.Error()
+	}
+
+	return spreadsheetID
+}
+
+// Returns the ID of the working spreadsheet for static tables
+func getSpreadsheetIDStatic(ctx context.Context, p *plugin.QueryData) string {
+	spreadsheetCacheKey := "googlesheets.spreadsheetID"
+	if cachedData, ok := p.ConnectionCache.Get(ctx, spreadsheetCacheKey); ok {
+		return cachedData.(string)
+	}
+
+	googleSheetsConfig := GetConfig(p.Connection)
+
+	var spreadsheetID string
+	if googleSheetsConfig.SpreadsheetId != nil {
+		spreadsheetID = *googleSheetsConfig.SpreadsheetId
+	}
+
+	err := p.ConnectionCache.Set(ctx, spreadsheetCacheKey, spreadsheetID)
+
+	if err != nil {
+		plugin.Logger(ctx).Error("getSpreadsheetID", "connection set error", err)
+		return err.Error()
+	}
 
 	return spreadsheetID
 }
@@ -47,7 +75,7 @@ func pathOrContents(poc string) (string, error) {
 
 	// Check for valid file path
 	if _, err := os.Stat(path); err == nil {
-		contents, err := ioutil.ReadFile(path)
+		contents, err := os.ReadFile(path)
 		if err != nil {
 			return string(contents), err
 		}
