@@ -3,10 +3,10 @@ package googlesheets
 import (
 	"context"
 
-	"github.com/turbot/steampipe-plugin-sdk/v3/plugin"
+	"github.com/turbot/steampipe-plugin-sdk/v5/plugin"
 )
 
-func listSpreadsheetWithPath(ctx context.Context, p *plugin.Plugin, sheetName string) func(ctx context.Context, d *plugin.QueryData, _ *plugin.HydrateData) (interface{}, error) {
+func listSpreadsheetWithPath(ctx context.Context, p *plugin.TableMapData, sheetName string) func(ctx context.Context, d *plugin.QueryData, _ *plugin.HydrateData) (interface{}, error) {
 	return func(ctx context.Context, d *plugin.QueryData, _ *plugin.HydrateData) (interface{}, error) {
 		// Get spreadsheet details
 		spreadsheetData, err := getSpreadsheetData(ctx, p, []string{sheetName})
